requirements.txt
httpx==0.23.3
pillow==9.0.1
UnityPy==1.10.0
argparse==1.4.0